Fix startup crash at fm_directory_view_request_thumbnails()
This fixes a bug that would make pantheon-files crash during startup when multiple instances were launched at once. Please note though that this only fixes the crash, and not the root problem; there appears to be a race condition between the directory-view requesting cached thumbnails and directory loading. Fixing that is a bigger task worth its own revision.