FileOperations: Re-escape uris from Clipboard or DNDHandler before use in fileoperations (#2616)
Fixes #2615

After #2268 files being copied or dragged have their name sanitized before being given to the recipient so that they appear correctly when pasted/dropped as text. However, they were not being re-escaped when used in a file operation and this caused the operation to fail for some files. This PR fixes that.

After this PR, copy/dragging a file should work both to a text document or terminal and to another folder in Files (or Dolphin or Nautilus at least).

Note that Dolphin does not unescape filenames when transferred as text but Nautilus does.

Co-authored-by: Leo <lenemter@gmail.com>